+++ pom.xml
@@ -28,7 +28,8 @@
         <module>scaledge-content</module>
         <module>scaledge-users</module>
         <module>scaledge-gateway</module>
-        <module>scaledge-db-docker</module>
+        <!--<module>scaledge-db-docker</module>-->
+        <module>scaledge-common</module>
 
     </modules>
     <dependencies>
