+++ scaledge-data/src/main/java/com/kgcorner/scaledgedata/models/BaseModel.java
@@ -1,4 +1,4 @@
-package com.kgcorner.dto;
+package com.kgcorner.scaledgedata.models;
 
 import org.springframework.data.annotation.Id;
